+++ packages/apptive_grid_core/lib/model/form/component/form_component.dart
@@ -85,6 +85,8 @@ abstract class FormComponent<T extends DataEntity> {
         return GeolocationFormComponent.fromJson(json);
       case DataType.multiCrossReference:
         return MultiCrossReferenceFormComponent.fromJson(json, properties);
+      case DataType.userReference:
+        return UserReferenceFormComponent.fromJson(json);
     }
   }
 }