+++ packages/apptive_grid_form/lib/widgets/form_widget/attachment_form_widget.dart
@@ -31,6 +31,7 @@ class _AttachmentFormWidgetState extends State<AttachmentFormWidget> {
           }
         },
         autovalidateMode: AutovalidateMode.onUserInteraction,
+        initialValue: widget.component.data,
         builder: (formState) {
           return InputDecorator(
             decoration: InputDecoration(