+++ packages/apptive_grid_form/lib/widgets/form_widget/attachment_manager.dart
@@ -25,5 +25,6 @@ class AttachmentManager {
     } else if (actionType == AttachmentActionType.add) {
       formData?.attachmentActions.remove(attachment);
     }
+    return null;
   }
 }