+++ packages/apptive_grid_form/lib/widgets/form_widget/enum_collection_form_widget.dart
@@ -30,6 +30,7 @@ class _EnumCollectionFormWidgetState extends State<EnumCollectionFormWidget> {
         }
       },
       autovalidateMode: AutovalidateMode.onUserInteraction,
+      initialValue: widget.component.data,
       builder: (formState) {
         return InputDecorator(
           decoration: InputDecoration(