+++ packages/apptive_grid_form/CHANGELOG.md
@@ -7,13 +7,25 @@
 ### New Data Types
 * `DataType.attachment`
 * `DataType.enumCollection`
-* `Data.multiCrossReference`
+* `DataType.multiCrossReference`
+* `DataType.createdBy`
 
 ### Bugfixes
 * Fixed Multiline Text FormWidgets
+* Fix required validation not working for all types if Form has initial data
+
+## [0.9.0-alpha.10]
+* Use stable version of `apptive_grid_core`
+
+## [0.9.0-alpha.9]
+* Add Support for `DataType.createdBy`
+* Fix required validation not working for all types if Form has initial data
+
+## [0.9.0-alpha.8]
+* Show Loading Indicator while submitting a Form
 
 ## [0.9.0-alpha.7]
-* Add proper support for [DataType.multiCrossReference] in Forms
+* Add proper support for `DataType.multiCrossReference` in Forms
 
 ## [0.9.0-alpha.6]
 * Add support for `DataType.multiCrossReference`
@@ -41,7 +53,7 @@
 * **BREAKING CHANGE** Requires minSdkVersion for Android to be at least 18
 
 ## [0.7.3+1]
-* Use Locale for DateTime Formats in [DateFormWidget] and [DateTimeFormWidget]
+* Use Locale for DateTime Formats in `DateFormWidget` and `DateTimeFormWidget`
 
 ## [0.7.3]
 * Add missing translations
@@ -54,7 +66,7 @@
 * Adding translations to Forms. Right now English and German are supported
 
 ## [0.7.1]
-* Adding a way to get [currentData] of a FormWidget
+* Adding a way to get `currentData` of a FormWidget
 
 ## [0.7.0]
 * **BREAKING CHANGE** Added `FormData` to `onActionSuccess` callback to receive the data that was successfully send